readd blog changes
