+++ app/blog/_assets/categoriesAndAuthors/categories.js
@@ -1,0 +1,30 @@
+import slugs from "./slugs";
+
+const { categorySlugs } = slugs;
+
+const categories = [
+  {
+    // The slug to use in the URL, from the categorySlugs object above.
+    slug: categorySlugs.feature,
+    // The title to display the category title (h1), the category badge, the category filter, and more. Less than 60 characters.
+    title: "New Features",
+    // A short version of the title above, display in small components like badges. 1 or 2 words
+    titleShort: "Features",
+    // The description of the category to display in the category page. Up to 160 characters.
+    description:
+      "Here are the latest features we've added to ShipFast. I'm constantly improving our product to help you ship faster.",
+    // A short version of the description above, only displayed in the <Header /> on mobile. Up to 60 characters.
+    descriptionShort: "Latest features added to ShipFast.",
+  },
+  {
+    slug: categorySlugs.tutorial,
+    title: "How Tos & Tutorials",
+    titleShort: "Tutorials",
+    description:
+      "Learn how to use ShipFast with these step-by-step tutorials. I'll show you how to ship faster and save time.",
+    descriptionShort:
+      "Learn how to use ShipFast with these step-by-step tutorials.",
+  },
+];
+
+export default categories;
