+++ app/blog/_assets/categoriesAndAuthors/slugs.js
@@ -1,0 +1,11 @@
+const slugs = {
+  categorySlugs: {
+    feature: "feature",
+    tutorial: "tutorial",
+  },
+  authorSlugs: {
+    marc: "marc"
+  }
+};
+
+export default slugs;
