+++ app/blog/page.js
@@ -5,9 +5,9 @@ import config from "@/config";
 import { getSEOTags } from "@/libs/seo";
 
 export const metadata = getSEOTags({
-  title: `${config.appName} Blog | Stripe Chargeback Protection`,
+  title: `${config.appName} Blog | Optimisez l'environnement de travail`,
   description:
-    "Learn how to prevent chargebacks, how to accept payments online, and keep your Stripe account in good standing",
+    "Découvrez comment rendre optimal l'environnement de travail de vos équipes",
   canonicalUrlRelative: "/blog",
 });
 
@@ -22,8 +22,7 @@ export default async function Blog() {
           The {config.appName} Blog
         </h1>
         <p className="text-lg opacity-80 leading-relaxed">
-          Learn how to ship your startup in days, not weeks. And get the latest
-          updates about the boilerplate
+          Découvrez comment rendre optimal l'environnement de travail de vos équipes
         </p>
       </section>
